+++ .github/workflows/python-app.yml
@@ -41,5 +41,8 @@ jobs:
     - name: Lint with black
       run: poetry run black . --check
     - name: Test with pytest
-      run: poetry run pytest
+      run: |
+      poetry run pytest
+      cd app
+      python -m manage test route_rangers_api.tests
