+++ app/scripts/extract_scheduled_gtfs.py
@@ -7,8 +7,6 @@ import gtfs_kit as gk
 import re
 from datetime import datetime
 
-import pdb
-
 from django.contrib.gis.geos import GEOSGeometry, LineString, Point, MultiLineString
 from route_rangers_api.models import TransitStation, TransitRoute, StationRouteRelation
 from django.db.utils import IntegrityError
